fix: replace Float with BigDecimal for currency precision (#21)
## Summary

- Replaces `Float` (7 significant digits) with `BigDecimal` throughout
the entire currency value pipeline — parse, cache, calculate, and
display
- All 10 adapters (JSON + XML) now parse values as `BigDecimal` instead
of `toFloat()`. JSON adapters read via `BigDecimal(reader.nextString())`
to preserve exact decimal precision from the source (never through
IEEE-754 double)
- SharedPreferences cache switches from `putFloat` to
`putString(toPlainString())`; any stale `Float` entries from old
installs are skipped and trigger a fresh API fetch
- All arithmetic in `MainViewModel` uses `BigDecimal.divide(…,
MathContext.DECIMAL128)` and `multiply` — no more precision loss in
`baseValue / baseRate * destinationRate`
- Info footer uses `toHumanReadableNumber` instead of
`String.format("%f", …)` to avoid silent downcast to double
- `TimelineViewModel` avg/min/max/difference all use BigDecimal
- `TextUtils.Float.toHumanReadableNumber` →
`BigDecimal.toHumanReadableNumber`
- `MathUtils.calculateDifference` and `getSignificantDecimalPlaces` use
BigDecimal
- Fee is stored as `String` (`toPlainString`) under a new `_fee_str`
key, with one-time migration from the legacy `_fee` Float key on first
read
- `ChartAdapter.getY` / `getBaseLine` convert to `Float` only at the
`SparkAdapter` boundary (library requirement)

Closes #7 (upstream)

## Test plan
- [ ] CI build passes
- [ ] Main conversion (e.g. 1 USD → EUR) shows correct result without
floating-point artifacts
- [ ] Fee toggle still works; existing fee setting is migrated on
upgrade (legacy `_fee` Float → new `_fee_str` String)
- [ ] Timeline chart renders, avg/min/max stats display correctly
- [ ] Conversion preview in currency picker shows correct values
- [ ] First launch after upgrade fetches fresh rates (old Float cache is
discarded gracefully)


---------

+++ app/src/main/kotlin/de/salomax/currencies/viewmodel/preference/PreferenceViewModel.kt
@@ -14,6 +14,7 @@ import de.salomax.currencies.repository.Database
 import de.salomax.currencies.repository.ExchangeRatesRepository
 import de.salomax.currencies.view.main.MainActivity
 import de.salomax.currencies.view.preference.PreferenceActivity
+import java.math.BigDecimal
 
 class PreferenceViewModel(private val app: Application) : AndroidViewModel(app) {
 
@@ -115,11 +116,11 @@ class PreferenceViewModel(private val app: Application) : AndroidViewModel(app) 
         return x || y
     }
 
-    fun setFee(fee: Float) {
+    fun setFee(fee: BigDecimal) {
         Database(app).setFee(fee)
     }
 
-    fun getFee(): LiveData<Float> {
+    fun getFee(): LiveData<BigDecimal> {
         return Database(app).getFee()
     }
